+++ src/strtoIf.c
@@ -34,7 +34,7 @@ THIS SOFTWARE.
 int strtoIf(const char *s, char **sp, float *f0, float *f1)
 {
 	static FPI fpi = {24, 1 - 127 - 24 + 1, 254 - 127 - 24 + 1, 1, SI};
-	Long exp[2];
+	int32_t exp[2];
 	Bigint* B[2];
 	int k;
 	int rv[2];
@@ -42,17 +42,17 @@ int strtoIf(const char *s, char **sp, float *f0, float *f1)
 	B[0] = Balloc(0);
 	B[0]->wds = 1;
 	k = strtoIg(s, sp, &fpi, exp, B, rv);
-	ULtof((ULong*)f0, B[0]->x, exp[0], rv[0]);
+	ULtof((uint32_t*)f0, B[0]->x, exp[0], rv[0]);
 	Bfree(B[0]);
 
 	if(B[1])
 	{
-		ULtof((ULong*)f1, B[1]->x, exp[1], rv[1]);
+		ULtof((uint32_t*)f1, B[1]->x, exp[1], rv[1]);
 		Bfree(B[1]);
 	}
 	else
 	{
-		*(ULong*)f1 = *(ULong*)f0;
+		*(uint32_t*)f1 = *(uint32_t*)f0;
 	}
 
 	return k;
